+++ CHANGELOG.md
@@ -21,6 +21,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 
+- (**breaking change**) `Iban.isValid(String)` or `IbanCheckDigit.validate(String)` return `false` instead of raising an
+  `IllegalArgumentException` with invalid IBAN check digit (e.g. `00`, `01`, or `99`).
+
 ### Deprecated
 
 ### Removed